--- apps/client/src/Features/Product/view/ListEmployee/index.js
@@ -1,1 +1,0 @@
-export { default } from "./ListEmployee";
