+++ apps/client/src/Features/Product/view/ListProduct/index.js
@@ -1,0 +1,1 @@
+export { default } from "./ListProduct";
